refactor: adjust deployer and run manager
