+++ pkg/deployer/terraform/deployer.go
@@ -2,79 +2,26 @@ package terraform
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"regexp"
-	"strings"
 
-	"github.com/hashicorp/terraform-config-inspect/tfconfig"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/kubernetes"
 
-	"github.com/seal-io/walrus/pkg/auths"
-	"github.com/seal-io/walrus/pkg/auths/session"
-	runbus "github.com/seal-io/walrus/pkg/bus/resourcerun"
-	"github.com/seal-io/walrus/pkg/dao"
 	"github.com/seal-io/walrus/pkg/dao/model"
-	"github.com/seal-io/walrus/pkg/dao/model/connector"
-	"github.com/seal-io/walrus/pkg/dao/model/environment"
-	"github.com/seal-io/walrus/pkg/dao/model/environmentconnectorrelationship"
-	"github.com/seal-io/walrus/pkg/dao/model/predicate"
-	"github.com/seal-io/walrus/pkg/dao/model/project"
-	"github.com/seal-io/walrus/pkg/dao/model/resource"
-	"github.com/seal-io/walrus/pkg/dao/model/resourcedefinitionmatchingrule"
-	"github.com/seal-io/walrus/pkg/dao/model/resourcerun"
-	"github.com/seal-io/walrus/pkg/dao/model/subject"
-	"github.com/seal-io/walrus/pkg/dao/model/template"
-	"github.com/seal-io/walrus/pkg/dao/model/templateversion"
 	"github.com/seal-io/walrus/pkg/dao/types"
-	"github.com/seal-io/walrus/pkg/dao/types/crypto"
-	"github.com/seal-io/walrus/pkg/dao/types/object"
 	"github.com/seal-io/walrus/pkg/dao/types/status"
 	deptypes "github.com/seal-io/walrus/pkg/deployer/types"
-	pkgenv "github.com/seal-io/walrus/pkg/environment"
-	opk8s "github.com/seal-io/walrus/pkg/operator/k8s"
-	pkgresource "github.com/seal-io/walrus/pkg/resource"
-	"github.com/seal-io/walrus/pkg/servervars"
+	pkgrun "github.com/seal-io/walrus/pkg/resourcerun"
+	runconfig "github.com/seal-io/walrus/pkg/resourcerun/config"
 	"github.com/seal-io/walrus/pkg/settings"
-	"github.com/seal-io/walrus/pkg/templates/openapi"
-	"github.com/seal-io/walrus/pkg/templates/translator"
-	"github.com/seal-io/walrus/pkg/terraform/config"
-	"github.com/seal-io/walrus/pkg/terraform/parser"
-	"github.com/seal-io/walrus/pkg/terraform/util"
-	"github.com/seal-io/walrus/utils/json"
 	"github.com/seal-io/walrus/utils/log"
-	"github.com/seal-io/walrus/utils/pointer"
 )
-
-// DeployerType the type of deployer.
-const DeployerType = types.DeployerTypeTF
-
-const (
-	// _backendAPI the API path to terraform deploy backend.
-	// Terraform will get and update deployment states from this API.
-	_backendAPI = "/v1/projects/%s/environments/%s/resources/%s/runs/%s/terraform-states"
-
-	// _variablePrefix the prefix of the variable name.
-	_variablePrefix = "_walrus_var_"
-
-	// _resourcePrefix the prefix of the resource output name.
-	_resourcePrefix = "_walrus_res_"
-)
-
-// _interpolationReg is the regular expression for matching non-reference or non-variable expressions.
-// Reference: https://developer.hashicorp.com/terraform/language/expressions/strings#escape-sequences-1
-// To handle escape sequences, ${xxx} is converted to $${xxx}.
-// If there are more than two consecutive $ symbols, like $${xxx}, they are further converted to $$${xxx}.
-// During Terraform processing, $${} is ultimately transformed back to ${},
-// this interpolation is used to ensuring a WYSIWYG user experience.
-var _interpolationReg = regexp.MustCompile(`\$\{((var\.)?([^.}]+)(?:\.([^.}]+))?)[^\}]*\}`)
 
 // Deployer terraform deployer to deploy the resource.
 type Deployer struct {
-	logger    log.Logger
-	clientSet *kubernetes.Clientset
+	logger          log.Logger
+	clientSet       *kubernetes.Clientset
+	runConfigLoader runconfig.Loader
 }
 
 func NewDeployer(_ context.Context, opts deptypes.CreateOptions) (deptypes.Deployer, error) {
@@ -84,13 +31,14 @@ func NewDeployer(_ context.Context, opts deptypes.CreateOptions) (deptypes.Deplo
 	}
 
 	return &Deployer{
-		clientSet: clientSet,
-		logger:    log.WithName("deployer").WithName("tf"),
+		clientSet:       clientSet,
+		logger:          log.WithName("deployer").WithName("tf"),
+		runConfigLoader: runconfig.NewInputLoader(types.DeployerTypeTF),
 	}, nil
 }
 
 func (d Deployer) Type() deptypes.Type {
-	return DeployerType
+	return types.DeployerTypeTF
 }
 
 // Apply creates a new resource run by the given resource,
@@ -101,10 +49,11 @@ func (d Deployer) Apply(
 	resource *model.Resource,
 	opts deptypes.ApplyOptions,
 ) (err error) {
-	run, err := d.createRun(ctx, mc, createRunOptions{
+	run, err := pkgrun.Create(ctx, mc, pkgrun.CreateOptions{
 		ResourceID:    resource.ID,
 		ChangeComment: resource.ChangeComment,
-		JobType:       JobTypeApply,
+		DeployerType:  types.DeployerTypeTF,
+		JobType:       types.RunJobTypeApply,
 	})
 	if err != nil {
 		return err
@@ -119,11 +68,11 @@ func (d Deployer) Apply(
 		status.ResourceRunStatusReady.False(run, err.Error())
 
 		// Report to resource run.
-		_ = d.updateRunStatus(ctx, mc, run)
+		_ = pkgrun.UpdateStatus(ctx, mc, run)
 	}()
 
 	return d.createK8sJob(ctx, mc, createK8sJobOptions{
-		Type:        JobTypeApply,
+		Type:        types.RunJobTypeApply,
 		ResourceRun: run,
 	})
 }
@@ -136,9 +85,10 @@ func (d Deployer) Destroy(
 	resource *model.Resource,
 	opts deptypes.DestroyOptions,
 ) (err error) {
-	run, err := d.createRun(ctx, mc, createRunOptions{
-		ResourceID: resource.ID,
-		JobType:    JobTypeDestroy,
+	run, err := pkgrun.Create(ctx, mc, pkgrun.CreateOptions{
+		ResourceID:   resource.ID,
+		DeployerType: types.DeployerTypeTF,
+		JobType:      types.RunJobTypeDestroy,
 	})
 	if err != nil {
 		return err
@@ -153,11 +103,11 @@ func (d Deployer) Destroy(
 		status.ResourceRunStatusReady.False(run, err.Error())
 
 		// Report to resource run.
-		_ = d.updateRunStatus(ctx, mc, run)
+		_ = pkgrun.UpdateStatus(ctx, mc, run)
 	}()
 
 	return d.createK8sJob(ctx, mc, createK8sJobOptions{
-		Type:        JobTypeDestroy,
+		Type:        types.RunJobTypeDestroy,
 		ResourceRun: run,
 	})
 }
@@ -171,55 +121,12 @@ type createK8sJobOptions struct {
 
 // createK8sJob creates a k8s job to deploy, destroy or rollback the resource.
 func (d Deployer) createK8sJob(ctx context.Context, mc model.ClientSet, opts createK8sJobOptions) error {
-	run := opts.ResourceRun
-
-	connectors, err := d.getConnectors(ctx, mc, run.EnvironmentID)
-	if err != nil {
-		return err
-	}
-
-	proj, err := mc.Projects().Get(ctx, run.ProjectID)
-	if err != nil {
-		return err
-	}
-
-	env, err := dao.GetEnvironmentByID(ctx, mc, run.EnvironmentID)
-	if err != nil {
-		return err
-	}
-
-	res, err := mc.Resources().Get(ctx, run.ResourceID)
-	if err != nil {
-		return err
-	}
-
-	var subjectID object.ID
-
-	sj, _ := session.GetSubject(ctx)
-	if sj.ID != "" {
-		subjectID = sj.ID
-	} else {
-		subjectID, err = pkgresource.GetSubjectID(res)
-		if err != nil {
-			return err
-		}
-	}
-
-	if subjectID == "" {
-		return errors.New("subject id is empty")
-	}
-
 	// Prepare tfConfig for deployment.
-	secretOpts := createK8sSecretsOptions{
-		ResourceRun: opts.ResourceRun,
-		Connectors:  connectors,
-		SubjectID:   subjectID,
-		// Walrus Context.
-		Context: *types.NewContext().
-			SetProject(proj.ID, proj.Name).
-			SetEnvironment(env.ID, env.Name, pkgenv.GetManagedNamespaceName(env)).
-			SetResource(res.ID, res.Name),
+	secretOpts, err := runconfig.GetConfigLoaderOptions(ctx, mc, opts.ResourceRun, _secretMountPath)
+	if err != nil {
+		return err
 	}
+
 	if err = d.createK8sSecrets(ctx, mc, secretOpts); err != nil {
 		return err
 	}
@@ -236,7 +143,7 @@ func (d Deployer) createK8sJob(ctx context.Context, mc model.ClientSet, opts cre
 		return err
 	}
 
-	// Create deployment job.
+	// Create a deployment job.
 	jobOpts := JobCreateOptions{
 		Type:          opts.Type,
 		ResourceRunID: opts.ResourceRun.ID.String(),
@@ -299,56 +206,29 @@ func (d Deployer) getEnv(ctx context.Context, mc model.ClientSet) (env []corev1.
 	return env
 }
 
-func (d Deployer) updateRunStatus(ctx context.Context, mc model.ClientSet, ar *model.ResourceRun) error {
-	// Report to resource run.
-	ar.Status.SetSummary(status.WalkResourceRun(&ar.Status))
-
-	ar, err := mc.ResourceRuns().UpdateOne(ar).
-		SetStatus(ar.Status).
-		Save(ctx)
-	if err != nil {
-		return err
-	}
-
-	if err = runbus.Notify(ctx, mc, ar); err != nil {
-		d.logger.Error(err)
-		return err
-	}
-
-	return nil
-}
-
-type createK8sSecretsOptions struct {
-	ResourceRun *model.ResourceRun
-	Connectors  model.Connectors
-	SubjectID   object.ID
-	// Walrus Context.
-	Context types.Context
-}
-
 // createK8sSecrets creates the k8s secrets for deployment.
-func (d Deployer) createK8sSecrets(ctx context.Context, mc model.ClientSet, opts createK8sSecretsOptions) error {
+func (d Deployer) createK8sSecrets(ctx context.Context, mc model.ClientSet, opts *runconfig.LoaderOptions) error {
 	secretData := make(map[string][]byte)
 	// SecretName terraform tfConfig name.
 	secretName := _jobSecretPrefix + string(opts.ResourceRun.ID)
 
 	// Prepare terraform config files bytes for deployment.
-	terraformData, err := d.loadConfigsBytes(ctx, mc, opts)
+	inputConfigs, err := d.runConfigLoader.LoadAll(ctx, mc, opts)
 	if err != nil {
 		return err
 	}
 
-	for k, v := range terraformData {
+	for k, v := range inputConfigs {
 		secretData[k] = v
 	}
 
 	// Mount the provider configs(e.g. kubeconfig) to secret.
-	providerData, err := d.getProviderSecretData(opts.Connectors)
+	providerConfigs, err := d.runConfigLoader.LoadProviders(opts.Connectors)
 	if err != nil {
 		return err
 	}
 
-	for k, v := range providerData {
+	for k, v := range providerConfigs {
 		secretData[k] = v
 	}
 
@@ -358,682 +238,4 @@ func (d Deployer) createK8sSecrets(ctx context.Context, mc model.ClientSet, opts
 	}
 
 	return nil
-}
-
-type createRunOptions struct {
-	// ResourceID indicates the ID of resource which is for create the run.
-	ResourceID object.ID
-	// ChangeComment indicates the optional message of the run.
-	ChangeComment string
-	// JobType indicates the type of the job.
-	JobType string
-}
-
-// createRun creates a new resource run.
-// Get the latest run, and check it if it is running.
-// If not running, then apply the latest run.
-// If running, then wait for the latest run to be applied.
-func (d Deployer) createRun(
-	ctx context.Context,
-	mc model.ClientSet,
-	opts createRunOptions,
-) (*model.ResourceRun, error) {
-	// Validate if there is a running run.
-	prevEntity, err := mc.ResourceRuns().Query().
-		Where(resourcerun.And(
-			resourcerun.ResourceID(opts.ResourceID),
-			resourcerun.DeployerType(DeployerType))).
-		Order(model.Desc(resourcerun.FieldCreateTime)).
-		First(ctx)
-	if err != nil && !model.IsNotFound(err) {
-		return nil, err
-	}
-
-	if prevEntity != nil && status.ResourceRunStatusReady.IsUnknown(prevEntity) {
-		return nil, errors.New("deployment is running")
-	}
-
-	// Get the corresponding resource and template version.
-	res, err := mc.Resources().Query().
-		Where(resource.ID(opts.ResourceID)).
-		WithTemplate(func(tvq *model.TemplateVersionQuery) {
-			tvq.Select(
-				templateversion.FieldName,
-				templateversion.FieldVersion,
-				templateversion.FieldTemplateID)
-		}).
-		WithProject(func(pq *model.ProjectQuery) {
-			pq.Select(project.FieldName, project.FieldLabels)
-		}).
-		WithEnvironment(func(env *model.EnvironmentQuery) {
-			env.Select(environment.FieldLabels)
-			env.Select(environment.FieldName)
-			env.Select(environment.FieldType)
-		}).
-		WithResourceDefinitionMatchingRule(func(mrq *model.ResourceDefinitionMatchingRuleQuery) {
-			mrq.Select(
-				resourcedefinitionmatchingrule.FieldName,
-				resourcedefinitionmatchingrule.FieldAttributes,
-			).
-				WithTemplate(func(tvq *model.TemplateVersionQuery) {
-					tvq.Select(
-						templateversion.FieldID,
-						templateversion.FieldVersion,
-						templateversion.FieldName,
-					)
-				})
-		}).
-		Select(
-			resource.FieldID,
-			resource.FieldProjectID,
-			resource.FieldEnvironmentID,
-			resource.FieldType,
-			resource.FieldLabels,
-			resource.FieldAnnotations,
-			resource.FieldAttributes,
-			resource.FieldComputedAttributes).
-		Only(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var (
-		templateID                    object.ID
-		templateName, templateVersion string
-		attributes                    = res.Attributes
-		computedAttributes            = res.ComputedAttributes
-	)
-
-	switch {
-	case res.TemplateID != nil:
-		templateID = res.Edges.Template.TemplateID
-		templateName = res.Edges.Template.Name
-		templateVersion = res.Edges.Template.Version
-	case res.ResourceDefinitionMatchingRuleID != nil:
-		rule := res.Edges.ResourceDefinitionMatchingRule
-
-		templateName = rule.Edges.Template.Name
-		templateVersion = rule.Edges.Template.Version
-
-		templateID, err = mc.Templates().Query().
-			Where(
-				template.Name(templateName),
-				// Now we only support resource definition globally.
-				template.ProjectIDIsNil(),
-			).
-			OnlyID(ctx)
-		if err != nil {
-			return nil, err
-		}
-	default:
-		return nil, errors.New("missing template or resource definition")
-	}
-
-	var subjectID object.ID
-
-	s, _ := session.GetSubject(ctx)
-	if s.ID != "" {
-		subjectID = s.ID
-	} else {
-		subjectID, err = pkgresource.GetSubjectID(res)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	userSubject, err := mc.Subjects().Query().
-		Where(subject.ID(subjectID)).
-		Only(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	entity := &model.ResourceRun{
-		ProjectID:          res.ProjectID,
-		EnvironmentID:      res.EnvironmentID,
-		ResourceID:         res.ID,
-		TemplateID:         templateID,
-		TemplateName:       templateName,
-		TemplateVersion:    templateVersion,
-		Attributes:         attributes,
-		ComputedAttributes: computedAttributes,
-		DeployerType:       DeployerType,
-		CreatedBy:          userSubject.Name,
-		ChangeComment:      opts.ChangeComment,
-	}
-
-	status.ResourceRunStatusReady.Unknown(entity, "")
-	entity.Status.SetSummary(status.WalkResourceRun(&entity.Status))
-
-	// Inherit the output of previous run to create a new one.
-	if prevEntity != nil {
-		entity.Output = prevEntity.Output
-	}
-
-	switch {
-	case opts.JobType == JobTypeApply && entity.Output != "":
-		// Get required providers from the previous output after first deployment.
-		requiredProviders, err := d.getRequiredProviders(ctx, mc, opts.ResourceID, entity.Output)
-		if err != nil {
-			return nil, err
-		}
-		entity.PreviousRequiredProviders = requiredProviders
-	case opts.JobType == JobTypeDestroy && entity.Output != "":
-		if status.ResourceRunStatusReady.IsFalse(prevEntity) {
-			// Get required providers from the previous output after first deployment.
-			requiredProviders, err := d.getRequiredProviders(ctx, mc, opts.ResourceID, entity.Output)
-			if err != nil {
-				return nil, err
-			}
-			entity.PreviousRequiredProviders = requiredProviders
-		} else {
-			// Copy required providers from the previous run.
-			entity.PreviousRequiredProviders = prevEntity.PreviousRequiredProviders
-			// Reuse other fields from the previous run.
-			entity.TemplateID = prevEntity.TemplateID
-			entity.TemplateName = prevEntity.TemplateName
-			entity.TemplateVersion = prevEntity.TemplateVersion
-			entity.Attributes = prevEntity.Attributes
-			entity.ComputedAttributes = prevEntity.ComputedAttributes
-			entity.InputPlan = prevEntity.InputPlan
-		}
-	}
-
-	// Create run.
-	entity, err = mc.ResourceRuns().Create().
-		Set(entity).
-		Save(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return entity, nil
-}
-
-func (d Deployer) getRequiredProviders(
-	ctx context.Context,
-	mc model.ClientSet,
-	instanceID object.ID,
-	previousOutput string,
-) ([]types.ProviderRequirement, error) {
-	stateRequiredProviderSet := sets.NewString()
-
-	previousRequiredProviders, err := d.getPreviousRequiredProviders(ctx, mc, instanceID)
-	if err != nil {
-		return nil, err
-	}
-
-	stateRequiredProviders, err := parser.ParseStateProviders(previousOutput)
-	if err != nil {
-		return nil, err
-	}
-
-	stateRequiredProviderSet.Insert(stateRequiredProviders...)
-
-	requiredProviders := make([]types.ProviderRequirement, 0, len(previousRequiredProviders))
-
-	for _, p := range previousRequiredProviders {
-		if stateRequiredProviderSet.Has(p.Name) {
-			requiredProviders = append(requiredProviders, p)
-		}
-	}
-
-	return requiredProviders, nil
-}
-
-// loadConfigsBytes returns terraform main.tf and terraform.tfvars for deployment.
-func (d Deployer) loadConfigsBytes(
-	ctx context.Context,
-	mc model.ClientSet,
-	opts createK8sSecretsOptions,
-) (map[string][]byte, error) {
-	logger := log.WithName("deployer").WithName("tf")
-	// Prepare terraform tfConfig.
-	//  get module configs from resource run.
-	moduleConfig, providerRequirements, err := d.getModuleConfig(ctx, mc, opts)
-	if err != nil {
-		return nil, err
-	}
-	// Merge current and previous required providers.
-	providerRequirements = append(providerRequirements,
-		opts.ResourceRun.PreviousRequiredProviders...)
-
-	requiredProviders := make(map[string]*tfconfig.ProviderRequirement, 0)
-	for _, p := range providerRequirements {
-		if _, ok := requiredProviders[p.Name]; !ok {
-			requiredProviders[p.Name] = p.ProviderRequirement
-		} else {
-			logger.Warnf("duplicate provider requirement: %s", p.Name)
-		}
-	}
-
-	runOpts := RunOpts{
-		ResourceRun:   opts.ResourceRun,
-		ResourceName:  opts.Context.Resource.Name,
-		ProjectID:     opts.Context.Project.ID,
-		EnvironmentID: opts.Context.Environment.ID,
-	}
-	// Parse module attributes.
-	attrs, variables, dependencyOutputs, err := ParseModuleAttributes(
-		ctx,
-		mc,
-		moduleConfig.Attributes,
-		false,
-		runOpts,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	moduleConfig.Attributes = attrs
-
-	// Update output sensitive with variables.
-	wrapVariables, err := updateOutputWithVariables(variables, moduleConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	// Prepare address for terraform backend.
-	serverAddress, err := settings.ServeUrl.Value(ctx, mc)
-	if err != nil {
-		return nil, err
-	}
-
-	if serverAddress == "" {
-		return nil, errors.New("server address is empty")
-	}
-	address := fmt.Sprintf("%s%s", serverAddress,
-		fmt.Sprintf(_backendAPI,
-			opts.Context.Project.ID,
-			opts.Context.Environment.ID,
-			opts.Context.Resource.ID,
-			opts.ResourceRun.ID))
-
-	// Prepare API token for terraform backend.
-	const _1Day = 60 * 60 * 24
-
-	at, err := auths.CreateAccessToken(ctx,
-		mc, opts.SubjectID, types.TokenKindDeployment, string(opts.ResourceRun.ID), pointer.Int(_1Day))
-	if err != nil {
-		return nil, err
-	}
-
-	// Prepare terraform config files to be mounted to secret.
-	requiredProviderNames := sets.NewString()
-	for _, p := range providerRequirements {
-		requiredProviderNames = requiredProviderNames.Insert(p.Name)
-	}
-
-	secretOptionMaps := map[string]config.CreateOptions{
-		config.FileMain: {
-			TerraformOptions: &config.TerraformOptions{
-				Token:                at.AccessToken,
-				Address:              address,
-				SkipTLSVerify:        !servervars.TlsCertified.Get(),
-				ProviderRequirements: requiredProviders,
-			},
-			ProviderOptions: &config.ProviderOptions{
-				RequiredProviderNames: requiredProviderNames.List(),
-				Connectors:            opts.Connectors,
-				SecretMonthPath:       _secretMountPath,
-				ConnectorSeparator:    parser.ConnectorSeparator,
-			},
-			ModuleOptions: &config.ModuleOptions{
-				ModuleConfigs: []*config.ModuleConfig{moduleConfig},
-			},
-			VariableOptions: &config.VariableOptions{
-				VariablePrefix:    _variablePrefix,
-				ResourcePrefix:    _resourcePrefix,
-				Variables:         wrapVariables,
-				DependencyOutputs: dependencyOutputs,
-			},
-			OutputOptions: moduleConfig.Outputs,
-		},
-		config.FileVars: getVarConfigOptions(variables, dependencyOutputs),
-	}
-	secretMaps := make(map[string][]byte, 0)
-
-	for k, v := range secretOptionMaps {
-		secretMaps[k], err = config.CreateConfigToBytes(v)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Save input plan to resource run.
-	opts.ResourceRun.InputPlan = string(secretMaps[config.FileMain])
-	// If resource run does not inherit variables from cloned run,
-	// then save the parsed variables to resource run.
-	if len(opts.ResourceRun.Variables) == 0 {
-		variableMap := make(crypto.Map[string, string], len(variables))
-		for _, s := range variables {
-			variableMap[s.Name] = string(s.Value)
-		}
-		opts.ResourceRun.Variables = variableMap
-	}
-
-	run, err := mc.ResourceRuns().UpdateOne(opts.ResourceRun).
-		Set(opts.ResourceRun).
-		Save(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	if err = runbus.Notify(ctx, mc, run); err != nil {
-		return nil, err
-	}
-
-	return secretMaps, nil
-}
-
-// getProviderSecretData returns provider kubeconfig secret data mount into terraform container.
-func (d Deployer) getProviderSecretData(connectors model.Connectors) (map[string][]byte, error) {
-	secretData := make(map[string][]byte)
-
-	for _, c := range connectors {
-		if c.Type != types.ConnectorTypeKubernetes {
-			continue
-		}
-
-		_, s, err := opk8s.LoadApiConfig(*c)
-		if err != nil {
-			return nil, err
-		}
-
-		// NB(alex) the secret file name must be config + connector id to
-		// match with terraform provider in config convert.
-		secretFileName := util.GetK8sSecretName(c.ID.String())
-		secretData[secretFileName] = []byte(s)
-	}
-
-	return secretData, nil
-}
-
-// getModuleConfig returns module configs and required connectors to
-// get terraform module config block from resource run.
-func (d Deployer) getModuleConfig(
-	ctx context.Context,
-	mc model.ClientSet,
-	opts createK8sSecretsOptions,
-) (*config.ModuleConfig, []types.ProviderRequirement, error) {
-	var (
-		requiredProviders = make([]types.ProviderRequirement, 0)
-		predicates        = make([]predicate.TemplateVersion, 0)
-	)
-
-	predicates = append(predicates, templateversion.And(
-		templateversion.Version(opts.ResourceRun.TemplateVersion),
-		templateversion.TemplateID(opts.ResourceRun.TemplateID),
-	))
-
-	templateVersion, err := mc.TemplateVersions().
-		Query().
-		Select(
-			templateversion.FieldID,
-			templateversion.FieldTemplateID,
-			templateversion.FieldName,
-			templateversion.FieldVersion,
-			templateversion.FieldSource,
-			templateversion.FieldSchema,
-			templateversion.FieldUiSchema,
-		).
-		Where(templateversion.Or(predicates...)).
-		Only(ctx)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if len(templateVersion.Schema.RequiredProviders) != 0 {
-		requiredProviders = append(requiredProviders, templateVersion.Schema.RequiredProviders...)
-	}
-
-	moduleConfig, err := getModuleConfig(opts.ResourceRun, templateVersion, opts)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return moduleConfig, requiredProviders, err
-}
-
-func (d Deployer) getConnectors(
-	ctx context.Context,
-	mc model.ClientSet,
-	environmentID object.ID,
-) (model.Connectors, error) {
-	rs, err := mc.EnvironmentConnectorRelationships().Query().
-		Where(environmentconnectorrelationship.EnvironmentID(environmentID)).
-		WithConnector(func(cq *model.ConnectorQuery) {
-			cq.Select(
-				connector.FieldID,
-				connector.FieldName,
-				connector.FieldType,
-				connector.FieldCategory,
-				connector.FieldConfigVersion,
-				connector.FieldConfigData)
-		}).
-		All(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var cs model.Connectors
-	for i := range rs {
-		cs = append(cs, rs[i].Edges.Connector)
-	}
-
-	return cs, nil
-}
-
-// getPreviousRequiredProviders get previous succeed run required providers.
-// NB(alex): the previous run may be failed, the failed run may not contain required providers of states.
-func (d Deployer) getPreviousRequiredProviders(
-	ctx context.Context,
-	mc model.ClientSet,
-	resourceID object.ID,
-) ([]types.ProviderRequirement, error) {
-	prevRequiredProviders := make([]types.ProviderRequirement, 0)
-
-	entity, err := mc.ResourceRuns().Query().
-		Where(resourcerun.ResourceID(resourceID)).
-		Order(model.Desc(resourcerun.FieldCreateTime)).
-		First(ctx)
-	if err != nil && !model.IsNotFound(err) {
-		return nil, err
-	}
-
-	if entity == nil {
-		return prevRequiredProviders, nil
-	}
-
-	templateVersion, err := mc.TemplateVersions().Query().
-		Where(
-			templateversion.TemplateID(entity.TemplateID),
-			templateversion.Version(entity.TemplateVersion),
-		).
-		Only(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(templateVersion.Schema.RequiredProviders) != 0 {
-		prevRequiredProviders = append(prevRequiredProviders, templateVersion.Schema.RequiredProviders...)
-	}
-
-	return prevRequiredProviders, nil
-}
-
-func getVarConfigOptions(
-	variables model.Variables,
-	resourceOutputs map[string]parser.OutputState,
-) config.CreateOptions {
-	varsConfigOpts := config.CreateOptions{
-		Attributes: map[string]any{},
-	}
-
-	for _, v := range variables {
-		varsConfigOpts.Attributes[_variablePrefix+v.Name] = v.Value
-	}
-
-	// Setup resource outputs.
-	for n, v := range resourceOutputs {
-		varsConfigOpts.Attributes[_resourcePrefix+n] = v.Value
-	}
-
-	return varsConfigOpts
-}
-
-func getModuleConfig(
-	run *model.ResourceRun,
-	template *model.TemplateVersion,
-	opts createK8sSecretsOptions,
-) (*config.ModuleConfig, error) {
-	mc := &config.ModuleConfig{
-		Name:   opts.Context.Resource.Name,
-		Source: template.Source,
-	}
-
-	if template.Schema.IsEmpty() {
-		return mc, nil
-	}
-
-	mc.SchemaData = template.Schema.TemplateVersionSchemaData
-
-	if template.Schema.OpenAPISchema == nil ||
-		template.Schema.OpenAPISchema.Components == nil ||
-		template.Schema.OpenAPISchema.Components.Schemas == nil {
-		return mc, nil
-	}
-
-	// Variables.
-	var (
-		variablesSchema    = template.Schema.VariableSchema()
-		outputsSchemas     = template.Schema.OutputSchema()
-		sensitiveVariables = sets.Set[string]{}
-	)
-
-	if variablesSchema != nil {
-		attrs, err := translator.ToGoTypeValues(run.ComputedAttributes, *variablesSchema)
-		if err != nil {
-			return nil, err
-		}
-
-		mc.Attributes = attrs
-
-		for n, v := range variablesSchema.Properties {
-			// Add sensitive from schema variable.
-			if v.Value.WriteOnly {
-				sensitiveVariables.Insert(fmt.Sprintf(`var\.%s`, n))
-			}
-
-			if n == types.WalrusContextVariableName {
-				mc.Attributes[n] = opts.Context
-			}
-		}
-	}
-
-	// Outputs.
-	if outputsSchemas != nil {
-		sps := outputsSchemas.Properties
-		mc.Outputs = make([]config.Output, 0, len(sps))
-
-		sensitiveVariableRegex, err := matchAnyRegex(sensitiveVariables.UnsortedList())
-		if err != nil {
-			return nil, err
-		}
-
-		for k, v := range sps {
-			origin := openapi.GetExtOriginal(v.Value.Extensions)
-			co := config.Output{
-				Sensitive:    v.Value.WriteOnly,
-				Name:         k,
-				ResourceName: opts.Context.Resource.Name,
-				Value:        origin.ValueExpression,
-			}
-
-			if !v.Value.WriteOnly {
-				// Update sensitive while output is from sensitive data, like secret.
-				if sensitiveVariables.Len() != 0 && sensitiveVariableRegex.Match(origin.ValueExpression) {
-					co.Sensitive = true
-				}
-			}
-
-			mc.Outputs = append(mc.Outputs, co)
-		}
-	}
-
-	return mc, nil
-}
-
-func updateOutputWithVariables(variables model.Variables, moduleConfig *config.ModuleConfig) (map[string]bool, error) {
-	var (
-		variableOpts         = make(map[string]bool)
-		encryptVariableNames = sets.NewString()
-	)
-
-	for _, s := range variables {
-		variableOpts[s.Name] = s.Sensitive
-
-		if s.Sensitive {
-			encryptVariableNames.Insert(_variablePrefix + s.Name)
-		}
-	}
-
-	if encryptVariableNames.Len() == 0 {
-		return variableOpts, nil
-	}
-
-	reg, err := matchAnyRegex(encryptVariableNames.UnsortedList())
-	if err != nil {
-		return nil, err
-	}
-
-	var shouldEncryptAttr []string
-
-	for k, v := range moduleConfig.Attributes {
-		b, err := json.Marshal(v)
-		if err != nil {
-			return nil, err
-		}
-
-		matches := reg.FindAllString(string(b), -1)
-		if len(matches) != 0 {
-			shouldEncryptAttr = append(shouldEncryptAttr, fmt.Sprintf(`var\.%s`, k))
-		}
-	}
-
-	// Outputs use encrypted variable should set to sensitive.
-	for i, v := range moduleConfig.Outputs {
-		if v.Sensitive {
-			continue
-		}
-
-		reg, err := matchAnyRegex(shouldEncryptAttr)
-		if err != nil {
-			return nil, err
-		}
-
-		if reg.MatchString(string(v.Value)) {
-			moduleConfig.Outputs[i].Sensitive = true
-		}
-	}
-
-	return variableOpts, nil
-}
-
-func matchAnyRegex(list []string) (*regexp.Regexp, error) {
-	var sb strings.Builder
-
-	sb.WriteString("(")
-
-	for i, v := range list {
-		sb.WriteString(v)
-
-		if i < len(list)-1 {
-			sb.WriteString("|")
-		}
-	}
-
-	sb.WriteString(")")
-
-	return regexp.Compile(sb.String())
 }
