+++ pkg/resourcerun/config/common.go
@@ -1,0 +1,108 @@
+package config
+
+import (
+	"fmt"
+	"regexp"
+	"strings"
+
+	"k8s.io/apimachinery/pkg/util/sets"
+
+	"github.com/seal-io/walrus/pkg/dao/model"
+	"github.com/seal-io/walrus/pkg/terraform/config"
+	"github.com/seal-io/walrus/pkg/terraform/parser"
+	"github.com/seal-io/walrus/utils/json"
+)
+
+func matchAnyRegex(list []string) (*regexp.Regexp, error) {
+	var sb strings.Builder
+
+	sb.WriteString("(")
+
+	for i, v := range list {
+		sb.WriteString(v)
+
+		if i < len(list)-1 {
+			sb.WriteString("|")
+		}
+	}
+
+	sb.WriteString(")")
+
+	return regexp.Compile(sb.String())
+}
+
+func updateOutputWithVariables(variables model.Variables, moduleConfig *config.ModuleConfig) (map[string]bool, error) {
+	var (
+		variableOpts         = make(map[string]bool)
+		encryptVariableNames = sets.NewString()
+	)
+
+	for _, s := range variables {
+		variableOpts[s.Name] = s.Sensitive
+
+		if s.Sensitive {
+			encryptVariableNames.Insert(_variablePrefix + s.Name)
+		}
+	}
+
+	if encryptVariableNames.Len() == 0 {
+		return variableOpts, nil
+	}
+
+	reg, err := matchAnyRegex(encryptVariableNames.UnsortedList())
+	if err != nil {
+		return nil, err
+	}
+
+	var shouldEncryptAttr []string
+
+	for k, v := range moduleConfig.Attributes {
+		b, err := json.Marshal(v)
+		if err != nil {
+			return nil, err
+		}
+
+		matches := reg.FindAllString(string(b), -1)
+		if len(matches) != 0 {
+			shouldEncryptAttr = append(shouldEncryptAttr, fmt.Sprintf(`var\.%s`, k))
+		}
+	}
+
+	// Outputs use encrypted variable should set to sensitive.
+	for i, v := range moduleConfig.Outputs {
+		if v.Sensitive {
+			continue
+		}
+
+		reg, err := matchAnyRegex(shouldEncryptAttr)
+		if err != nil {
+			return nil, err
+		}
+
+		if reg.MatchString(string(v.Value)) {
+			moduleConfig.Outputs[i].Sensitive = true
+		}
+	}
+
+	return variableOpts, nil
+}
+
+func getVarConfigOptions(
+	variables model.Variables,
+	resourceOutputs map[string]parser.OutputState,
+) config.CreateOptions {
+	varsConfigOpts := config.CreateOptions{
+		Attributes: map[string]any{},
+	}
+
+	for _, v := range variables {
+		varsConfigOpts.Attributes[_variablePrefix+v.Name] = v.Value
+	}
+
+	// Setup resource outputs.
+	for n, v := range resourceOutputs {
+		varsConfigOpts.Attributes[_resourcePrefix+n] = v.Value
+	}
+
+	return varsConfigOpts
+}
