+++ pkg/terraform/util/helper.go
@@ -7,5 +7,5 @@ import (
 // GetK8sSecretName returns the secret config name for the given connector.
 // Used for kubernetes connector. Or other connectors which need to store the kubeconfig in secret.
 func GetK8sSecretName(connectorID string) string {
-	return fmt.Sprintf("config%s", connectorID)
+	return fmt.Sprintf("config-%s", connectorID)
 }
